+++ app/views/__init__.py
@@ -1,0 +1,3 @@
+from .atlas_views import *
+from .entry_views import *
+from .misc_views import *
